pyproject.toml
[tool.poetry]
name = "scan"
version = "0.1.0"
description = ""
authors = ["comtext <comtext2022@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
autopep8 = "^2.0.0"
"ruamel.yaml" = "^0.17.21"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
